feat: tornando os usuários proprietários de seus dados fix: alterando o relacionamento entre o model Grafico e User (OneToOne) fix: reorganizando o sistema de filtragem ao listar (pavilhões, salas e ares)

+++ ac_mm_system/website/views.py
@@ -1,6 +1,8 @@
 # Create your views here.
 from django.contrib.auth.decorators import login_required
-from django.shortcuts import render, redirect
+from django.db.models.functions import Lower
+from django.http import Http404
+from django.shortcuts import render, redirect, get_object_or_404
 
 from django.contrib import messages
 
@@ -15,25 +17,28 @@ from chartjs.views.lines import BaseLineChartView
 
 @login_required
 def listar_pavilhoes(request):
+    pavilhoes = Pavilhao.objects.filter(usuario=request.user).order_by(Lower('nome'))
+
     context = {
-        'pavilhoes': Pavilhao.objects.all()  # permite exibir todos os objetos 'pavilhão' que foram criados
+        'pavilhoes': pavilhoes
     }
     return render(request, 'listar_pavilhoes.html', context)
 
 
 @login_required
 def listar_salas(request):
-    pavilhoes = Pavilhao.objects.all()
+    pavilhoes = Pavilhao.objects.filter(usuario=request.user).order_by(Lower('nome'))
+
     filtrarpavilhao = None
 
     if request.method == 'POST':
         filtrarpavilhao = request.POST.get('pavilhao')
         if filtrarpavilhao:
-            salas = Sala.objects.filter(pavilhao__uuid=filtrarpavilhao)
+            salas = Sala.objects.filter(pavilhao__uuid=filtrarpavilhao).order_by(Lower('nome'))
         else:
-            salas = Sala.objects.all()
+            salas = Sala.objects.filter(pavilhao__usuario=request.user).order_by(Lower('pavilhao__nome'), Lower('nome'))
     else:
-        salas = Sala.objects.all()
+        salas = Sala.objects.filter(pavilhao__usuario=request.user).order_by(Lower('pavilhao__nome'), Lower('nome'))
 
     context = {
         'salas': salas,
@@ -45,18 +50,25 @@ def listar_salas(request):
 
 @login_required
 def listar_ares(request):
-    pavilhoes = Pavilhao.objects.all()
-    salas = Sala.objects.all()
-    ares = ArCondicionado.objects.all()
+    pavilhoes = Pavilhao.objects.filter(usuario=request.user).order_by(Lower('nome'))
+    salas = Sala.objects.filter(pavilhao__usuario=request.user).order_by(Lower('nome'))
+    ares = ArCondicionado.objects.filter(sala__pavilhao__usuario=request.user)
 
     filtrarpavilhao = request.POST.get('pavilhao')
     filtrarsala = request.POST.get('sala')
 
     if filtrarpavilhao:
+        salas = salas.filter(pavilhao__uuid=filtrarpavilhao).order_by(Lower('nome'))
         ares = ares.filter(sala__pavilhao__uuid=filtrarpavilhao)
-        salas = salas.filter(pavilhao__uuid=filtrarpavilhao)  # Atualiza as salas conforme o pavilhão
+
     if filtrarsala:
         ares = ares.filter(sala__uuid=filtrarsala)
+
+    ares = ares.order_by(
+        Lower('sala__pavilhao__nome'),
+        Lower('sala__nome'),
+        Lower('nome')
+    )
 
     context = {
         'ares': ares,
@@ -71,8 +83,8 @@ def listar_ares(request):
 
 @login_required
 def listar_horarios(request):
-    pavilhoes = Pavilhao.objects.all()
-    salas = Sala.objects.all()
+    pavilhoes = Pavilhao.objects.filter(usuario=request.user).order_by(Lower('nome'))
+    salas = Sala.objects.filter(pavilhao__usuario=request.user).order_by(Lower('nome'))
     turnos = [
         ('Madrugada', 'Madrugada'),
         ('Matutino', 'Matutino'),
@@ -81,7 +93,8 @@ def listar_horarios(request):
     ]
 
     dias_da_semana = ["Segunda", "Terça", "Quarta", "Quinta", "Sexta", "Sábado", "Domingo"]
-    horas = sorted(set(horario.horario_inicio.hour for horario in Horario.objects.all()))
+    horas = sorted(
+        set(horario.horario_inicio.hour for horario in Horario.objects.filter(sala__pavilhao__usuario=request.user)))
 
     filtrarpavilhao = request.POST.get('pavilhao')
     filtrarturno = request.POST.get('turno')
@@ -123,13 +136,15 @@ def listar_horarios(request):
 @login_required
 def criar_pavilhao(request):
     # Validação e envio do formulário da sala
-    if str(request.method) == 'POST':
+    if request.method == 'POST':
         form = PavilhaoModelForm(request.POST)
         if form.is_valid():  # Verifica se os dados inseridos são válidos
-            form.save()  # Se os dados forem válidos ele irá salvar
+            novo_pavilhao = form.save(commit=False)
+            novo_pavilhao.usuario = request.user
+            novo_pavilhao.save()
             messages.success(request,
                              'Pavilhão criado com sucesso!')  # Exibe uma mensagem de sucesso, caso o pavilhão seja criado
-            # form = PavilhaoModelForm()  # Cria um novo formulário vazio após salvar os dados
+
             return redirect('website:listar_pavilhoes')
         else:
             messages.error(request,
@@ -145,16 +160,15 @@ def criar_pavilhao(request):
 @login_required
 def criar_sala(request):
     # Validação e envio do formulário da sala
-    if str(request.method) == 'POST':
-        form = SalaModelForm(request.POST)
+    if request.method == 'POST':
+        form = SalaModelForm(request.POST, usuario=request.user)
         if form.is_valid():  # Verifica se os dados inseridos são válidos
-            form.save()  # Se os dados forem válidos ele irá salvar
+            form.save()
             messages.success(request,
                              'Sala criada com sucesso!')  # Exibe uma mensagem de sucesso, caso a sala seja criada
-            # form = SalaModelForm()  # Cria um novo formulário vazio após salvar os dados
             return redirect('website:listar_salas')
     else:
-        form = SalaModelForm()
+        form = SalaModelForm(usuario=request.user)
     context = {
         'form': form  # Passa o formulário para o contexto do template
     }
@@ -164,15 +178,15 @@ def criar_sala(request):
 @login_required
 def criar_ar(request):
     # Validação e envio do formulário da sala
-    if str(request.method) == 'POST':
-        form = ArCondicionadoModelForm(request.POST)
+    if request.method == 'POST':
+        form = ArCondicionadoModelForm(request.POST, usuario=request.user)
         if form.is_valid():  # Verifica se os dados inseridos são validos
             form.save()  # Se os dados forem válidos, ele irá salvar
             messages.success(request,
                              'Ar-condicionado criado com sucesso!')
             return redirect('website:listar_ares')
     else:
-        form = ArCondicionadoModelForm()
+        form = ArCondicionadoModelForm(usuario=request.user)
     context = {
         'form': form  # Passa o formulário para o contexto do template
     }
@@ -182,8 +196,8 @@ def criar_ar(request):
 @login_required
 def criar_horario(request):
     # Validação e envio do formulário da sala
-    if str(request.method) == 'POST':
-        form = HorarioModelForm(request.POST)
+    if request.method == 'POST':
+        form = HorarioModelForm(request.POST, usuario=request.user)
         if form.is_valid():  # Verifica se os dados inseridos são válidos
             form.save()  # Se os dados forem válidos ele irá salvar
             messages.success(request,
@@ -194,7 +208,7 @@ def criar_horario(request):
             messages.error(request,
                            'Erro ao criar horário')  # Exibe uma mensagem de erro, caso os dados não sejam válidos
     else:
-        form = HorarioModelForm()
+        form = HorarioModelForm(usuario=request.user)
     context = {
         'form': form
     }
@@ -203,24 +217,33 @@ def criar_horario(request):
 
 @login_required
 def editar_salas(request, uuid):
-    sala = Sala.objects.get(uuid=uuid)
+    sala = get_object_or_404(Sala, uuid=uuid)
+    pavilhao = sala.pavilhao
+
+    if pavilhao.usuario != request.user:
+        raise Http404
 
     if request.method == 'POST':
-        form = SalaModelForm(request.POST, instance=sala)
+        form = SalaModelForm(request.POST, instance=sala, usuario=request.user)
         if form.is_valid():
             form.save()
             messages.success(request,
                              'Sala editada com sucesso!')
             return redirect('website:listar_salas')
     else:
-        form = SalaModelForm(instance=sala)
+        form = SalaModelForm(instance=sala, usuario=request.user)
     context = {'form': form, 'sala': sala}
     return render(request, 'criar_sala.html', context)
 
 
 @login_required
 def deletar_salas(request, uuid):
-    sala = Sala.objects.get(uuid=uuid)
+    sala = get_object_or_404(Sala, uuid=uuid)
+    pavilhao = sala.pavilhao
+
+    if pavilhao.usuario != request.user:
+        raise Http404
+
     if request.method == 'POST':
         sala.delete()
         messages.success(request,
@@ -234,7 +257,12 @@ def deletar_salas(request, uuid):
 
 @login_required
 def deletar_horarios(request, uuid):
-    horario = Horario.objects.get(uuid=uuid)
+    horario = get_object_or_404(Horario, uuid=uuid)
+    pavilhao = horario.sala.pavilhao
+
+    if pavilhao.usuario != request.user:
+        raise Http404
+
     if request.method == 'POST':
         horario.delete()
         messages.success(request,
@@ -248,24 +276,32 @@ def deletar_horarios(request, uuid):
 
 @login_required
 def editar_horarios(request, uuid):
-    horario = Horario.objects.get(uuid=uuid)
+    horario = get_object_or_404(Horario, uuid=uuid)
+    pavilhao = horario.sala.pavilhao
+
+    if pavilhao.usuario != request.user:
+        raise Http404
 
     if request.method == 'POST':
-        form = HorarioModelForm(request.POST, instance=horario)
+        form = HorarioModelForm(request.POST, instance=horario, usuario=request.user)
         if form.is_valid():
             form.save()
             messages.success(request,
                              'Horário editado com sucesso!')
             return redirect('website:listar_horarios')
     else:
-        form = HorarioModelForm(instance=horario)
+        form = HorarioModelForm(instance=horario, usuario=request.user)
     context = {'form': form, 'horario': horario}
     return render(request, 'criar_horario.html', context)
 
 
 @login_required
 def deletar_pavilhoes(request, uuid):
-    pavilhao = Pavilhao.objects.get(uuid=uuid)
+    pavilhao = get_object_or_404(Pavilhao, uuid=uuid)
+
+    if pavilhao.usuario != request.user:
+        raise Http404
+
     if request.method == 'POST':
         pavilhao.delete()
         messages.success(request,
@@ -279,7 +315,10 @@ def deletar_pavilhoes(request, uuid):
 
 @login_required
 def editar_pavilhoes(request, uuid):
-    pavilhao = Pavilhao.objects.get(uuid=uuid)
+    pavilhao = get_object_or_404(Pavilhao, uuid=uuid)
+
+    if pavilhao.usuario != request.user:
+        raise Http404
 
     if request.method == 'POST':
         form = PavilhaoModelForm(request.POST, instance=pavilhao)
@@ -296,24 +335,32 @@ def editar_pavilhoes(request, uuid):
 
 @login_required
 def editar_ares(request, uuid):
-    ar = ArCondicionado.objects.get(uuid=uuid)
+    ar = get_object_or_404(ArCondicionado, uuid=uuid)
+    pavilhao = ar.sala.pavilhao
+
+    if pavilhao.usuario != request.user:
+        raise Http404
 
     if request.method == 'POST':
-        form = ArCondicionadoModelForm(request.POST, instance=ar)
+        form = ArCondicionadoModelForm(request.POST, instance=ar, usuario=request.user)
         if form.is_valid():
             form.save()
             messages.success(request,
                              'Ar-condicionado editado com sucesso!')
             return redirect('website:listar_ares')
     else:
-        form = ArCondicionadoModelForm(instance=ar)
+        form = ArCondicionadoModelForm(instance=ar, usuario=request.user)
     context = {'form': form, 'ar': ar}
     return render(request, 'criar_ar.html', context)
 
 
 @login_required
 def deletar_ares(request, uuid):
-    ar = ArCondicionado.objects.get(uuid=uuid)
+    ar = get_object_or_404(ArCondicionado, uuid=uuid)
+    pavilhao = ar.sala.pavilhao
+
+    if pavilhao.usuario != request.user:
+        raise Http404
 
     if request.method == 'POST':
         ar.delete()
@@ -326,7 +373,11 @@ def deletar_ares(request, uuid):
 
 @login_required
 def ajustar_ar(request, uuid):
-    ar = ArCondicionado.objects.get(uuid=uuid)
+    ar = get_object_or_404(ArCondicionado, uuid=uuid)
+    pavilhao = ar.sala.pavilhao
+
+    if pavilhao.usuario != request.user:
+        raise Http404
 
     context = {
         'ar': ar
@@ -337,9 +388,13 @@ def ajustar_ar(request, uuid):
 
 @login_required
 def ajustar_sala(request, uuid):
-    sala = Sala.objects.get(uuid=uuid)
+    sala = get_object_or_404(Sala, uuid=uuid)
+    pavilhao = sala.pavilhao
     ares = ArCondicionado.objects.filter(sala=sala)
     ares_quantidade = ares.count()
+
+    if pavilhao.usuario != request.user:
+        raise Http404
 
     context = {
         'sala': sala,
@@ -352,8 +407,13 @@ def ajustar_sala(request, uuid):
 
 @login_required
 def ajustes_ares(request, uuid):
-    ar = ArCondicionado.objects.get(uuid=uuid)  # Obtém o ar-condicionado atualizado
-    sala = ar.sala  # Pega a sala correta
+    ar = get_object_or_404(ArCondicionado, uuid=uuid)
+    sala = ar.sala
+    pavilhao = sala.pavilhao
+
+    if pavilhao.usuario != request.user:
+        raise Http404
+
     if request.method == 'POST':
         comando = request.POST.get('comando')
         topico = sala.topico_mqtt
@@ -372,7 +432,11 @@ def ajustes_ares(request, uuid):
 
 @login_required
 def ajustes_salas(request, uuid):
-    sala = Sala.objects.get(uuid=uuid)  # Obtém a sala
+    sala = get_object_or_404(Sala, uuid=uuid)
+    pavilhao = sala.pavilhao
+
+    if pavilhao.usuario != request.user:
+        raise Http404
 
     if request.method == 'POST':
         comando = request.POST.get('comando')
@@ -389,14 +453,17 @@ def ajustes_salas(request, uuid):
 
         return redirect('website:ajustar_sala', uuid=uuid)
 
+
 @login_required
 def pagina_inicial(request):
-    # Buscar o primeiro objeto Grafico ou definir valor padrão
-    grafico = Grafico.objects.first()
+    usuario = request.user
+
+    # Buscar gráfico do usuário logado
+    grafico = Grafico.objects.filter(usuario=usuario).first()
     valor_kWh = grafico.valor_kWh if grafico else 0.0
 
-    # Obter todos os pavilhões
-    pavilhoes = list(Pavilhao.objects.all())
+    # Obter pavilhões do usuário
+    pavilhoes = list(Pavilhao.objects.filter(usuario=usuario))
 
     # Criar dicionário de gastos por pavilhão
     gasto_pav = {pav.nome: pav.consumo_total() * valor_kWh for pav in pavilhoes}
@@ -405,10 +472,10 @@ def pagina_inicial(request):
     consumo_total_geral = sum(pav.consumo_total() for pav in pavilhoes)
     gasto_total_geral = consumo_total_geral * valor_kWh
 
-    # Criar dicionário de dados do gráfico de pizza
+    # Dados para gráfico de pizza
     dados_pizza = {pav.nome: gasto_pav.get(pav.nome, 0) for pav in pavilhoes}
 
-    # Calcular consumo por dia da semana
+    # Consumo por dia da semana
     dias_semana = ["Segunda", "Terça", "Quarta", "Quinta", "Sexta", "Sábado", "Domingo"]
     consumo_diario = {dia: 0 for dia in dias_semana}
 
@@ -417,26 +484,22 @@ def pagina_inicial(request):
             for horario in sala.horarios.all():
                 dias = [d.strip(" []'") for d in horario.dias_da_semana.split(",")]
 
-                # Calcular o tempo de funcionamento diário em horas
                 inicio = horario.horario_inicio.hour + (horario.horario_inicio.minute / 60)
                 fim = horario.horario_fim.hour + (horario.horario_fim.minute / 60)
 
-                if fim < inicio:  # Caso o horário passe da meia-noite
+                if fim < inicio:
                     horas_uso = (24 - inicio) + fim
                 else:
                     horas_uso = fim - inicio
 
-                # Somar consumo de cada ar-condicionado da sala
                 for ac in sala.ares_condicionados.all():
-                    consumo_kWh = ac.potencia_kw * horas_uso  # Consumo diário desse AC
+                    consumo_kWh = ac.potencia_kw * horas_uso
                     for dia in dias:
                         if dia in consumo_diario:
                             consumo_diario[dia] += consumo_kWh
 
-    # Calcular o total da semana
     total_kWh_semana = sum(consumo_diario.values())
 
-    # Criar dicionário de dados para o gráfico de barras
     dados_barras = OrderedDict((dia, consumo_diario[dia]) for dia in dias_semana)
 
     context = {
@@ -452,14 +515,19 @@ def pagina_inicial(request):
 
 @login_required
 def editar_grafico(request):
-    grafico = Grafico.objects.first()
+    usuario = request.user
+    grafico, created = Grafico.objects.get_or_create(usuario=usuario)
+
     if request.method == 'POST':
         form = GraficoModelForm(request.POST, instance=grafico)
         if form.is_valid():
-            form.save()
+            novo_grafico = form.save(commit=False)
+            novo_grafico.usuario = usuario
+            novo_grafico.save()
             return redirect('website:pagina_inicial')
     else:
         form = GraficoModelForm(instance=grafico)
+
     context = {
         'form': form,
         'grafico': grafico
